Rename DataStorageOptimizer to DataStorageCompactor
